Refactor subjects app
