+++ src/apps/subjects/api/serializers/__init__.py
@@ -1,0 +1,1 @@
+from .subject_serializers import SubjectSerializer
