+++ src/apps/subjects/api/views/__init__.py
@@ -1,0 +1,1 @@
+from .subject_view_set import SubjectViewset
