+++ src/apps/subjects/api/views/subject_view_set.py
@@ -1,5 +1,5 @@
 from apps.shared.api.views import GenericModelViewSet
-from apps.subjects.subject_api.subject_serializers.subjectSerializer import (
+from apps.subjects.api.serializers import (
     SubjectSerializer,
 )
