+++ src/apps/subjects/models/__init__.py
@@ -1,0 +1,1 @@
+from .subject import Subject
